# src/docs/indexSwagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0",
    "title": "FIFO-Server-API",
    "description": "Api para o desafio do processo de seleção da FCamara"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
  "host": "localhost:3333",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/localization": {
      "get": {
        "tags": [
          "Rotas Publicas"
        ],
        "summary": "Consulta todos as localizacoes.",
        "operationId": "getLocalization",
        "description": "Esta operação tem por objetivo consultar todos as localizacoes",
        "responses": {
          "200": {
            "description": "Sucesso ao consultar todos os clientes.",
            "schema": {
              "$ref": "#/definitions/Localizations"
            }
          },
          "400": {
            "description": "Consulta inválida."
          },
          "401": {
            "description": "Consulta não autorizada."
          },
          "404": {
            "description": "Nenhuma localizacao encontrada para a consulta."
          },
          "500": {
            "description": "Erro interno ao tentar realizar consulta."
          }
        }
      }
    },
...
    "/localization_service/{id}": {
      "get": {
        "tags": [
          "Rotas Publicas"
        ],
        "summary": "Recupera uma lista de servicos da localizacao por id.",
        "description": "Esta operacao tem por objetivo recuperar uma lista de servicos da localizacao por id.",
        "operationId": "findLocalization",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "required": true,
            "description": "Numero do id da localizacao"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso ao consultar uma localizacao existente por id.",
            "schema": {
              "$ref": "#/definitions/LocalizationService"
            }
          },
          "400": {
            "description": "Id fornecido invalido."
          },
          "404": {
            "description": "Objeto não encontrado."
          },
          "500": {
            "description": "Erro interno ao tentar deletar."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Localizations": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Localization"
      }
    },
    "Localization": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "Id da localizacao."
        },
        "description": {
          "type": "string",
          "description": "Descrição da localizacao"
        },
        "name": {
          "type": "string",
          "description": "Nome da localizacao."
        },
        "notes": {
          "type": "string",
          "description": "Anotações da localizacao."
        }
      }
    },
...
    "LocalizationService": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "Id da localizacaoXservico"
        },
        "localization_id": {
          "type": "integer",
          "format": "int32",
          "description": "Id da localizacao"
        },
        "service_id": {
          "type": "integer",
          "format": "int32",
          "description": "Id do servico"
        }
      }
    },
...
  }
}